Search for dSYM files inside bundles
atosym previously expected the path to a dSYM
file, but Xcode actually produces dSYM bundles.
Make things easier for the user by automatically
looking for dSYM files inside bundles when the
passed in path is a directory. This spares the
user from having to specify the full path to
the actual dSYM file in the majority of cases.

See also:

  https://wincent.com/issues/1273

Signed-off-by: Wincent Colaiuta <win@wincent.com>